Implemented creation and login with database checks

+++ rest-api/src/main/java/net/cpollet/itinerants/ws/api/v1/data/LoginResponse.java
@@ -21,14 +21,14 @@ public class LoginResponse {
 
     private String token;
     private Set<String> roles;
-    private Integer personId;
+    private Long personId;
     private String result;
 
     private LoginResponse() {
         // nothing to do
     }
 
-    public LoginResponse(String token, Integer personId, Set<String> roles) {
+    public LoginResponse(String token, Long personId, Set<String> roles) {
         this.token = token;
         this.roles = roles;
         this.result = CODE_SUCCESS;
